Fix bugs and improve print messages
Fix a major bug that resulted in the storage orders loaded from the
files having empty basket. Also, fixed a bug that led to Exceptions
while parsing the year, month and day of the order date in the Analyze
Data class. Finally improved print messages and added \n characters in
order to display the messages more clearly.